+++ src/components/TextContainer/TextContainer.tsx
@@ -1,10 +1,18 @@
 import * as React from 'react';
+import {classNames, variationName} from '@shopify/react-utilities/styles';
+
 import * as styles from './TextContainer.scss';
 
+export type Spacing = 'tight' | 'loose';
 export interface Props {
+  spacing?: Spacing,
   children?: React.ReactNode,
 }
 
-export default function TextContainer({children}: Props) {
-  return <div className={styles.TextContainer}>{children}</div>;
+export default function TextContainer({spacing, children}: Props) {
+  const className = classNames(
+    styles.TextContainer,
+    spacing && styles[variationName('spacing', spacing)],
+  );
+  return <div className={className}>{children}</div>;
 }