+++ src/embedded/ResourcePicker/ResourcePicker.tsx
@@ -1,20 +1,9 @@
 import * as React from 'react';
 import {withEASDK, WithEASDKProps} from '../easdk';
+import {OpenOptions} from '../easdk/components/ResourcePicker';
 
-export interface SelectionResult {
-  products?: object[],
-  collections?: object[],
-}
-
-export interface Props {
-  title?: string,
+export interface Props extends OpenOptions {
   open: boolean,
-  products?: boolean,
-  collections?: boolean,
-  allowMultiple?: boolean,
-  showHidden?: boolean,
-  onCancel?(): void,
-  onSelection?(selection: SelectionResult): void,
 }
 
 export class ResourcePicker extends React.PureComponent<Props & WithEASDKProps, never> {